refactor: Rename package name of each tests
