+++ tools/integration-tests/tester/tests/common/main_test.go
@@ -1,4 +1,4 @@
-package autopeering
+package common
 
 import (
 	"os"
