+++ tools/integration-tests/tester/tests/drng/drng_test.go
@@ -1,4 +1,4 @@
-package autopeering
+package drng
 
 import (
 	"encoding/json"
