+++ tools/integration-tests/tester/tests/drng/main_test.go
@@ -1,4 +1,4 @@
-package autopeering
+package drng
 
 import (
 	"os"
