+++ tools/integration-tests/tester/tests/message/main_test.go
@@ -1,4 +1,4 @@
-package autopeering
+package message
 
 import (
 	"os"
