+++ tools/integration-tests/tester/tests/message/message_test.go
@@ -1,4 +1,4 @@
-package autopeering
+package message
 
 import (
 	"testing"
