+++ tools/integration-tests/tester/tests/value/main_test.go
@@ -1,4 +1,4 @@
-package autopeering
+package value
 
 import (
 	"os"
