+++ tools/integration-tests/tester/tests/value/value_test.go
@@ -1,4 +1,4 @@
-package autopeering
+package value
 
 import (
 	"testing"
